Reorganize the javascript code
