+++ js/src/Geonotebook.js
@@ -1,92 +1,11 @@
-import './css/styles.css';
-import _ from 'underscore';
-
-import * as jsonrpc from './jsonrpc';
-import GeoMap from './map';
-
-// This class stores a JSONRPC message and callbacks which are evaluated
-// once the Remote object recieves a 'resolve' call with the message's id.
-// This is initialized with a JSONRPC message and a function that takes a
-// message and sends it across some transport mechanism (e.g. Websocket).
-var ReplyCallback = function (msg, send_message) {
-  this.state = 'CREATED';
-  this.msg = msg;
-  this.send_message = send_message;
-};
-
-ReplyCallback.prototype.then = function (reply_handler, error_handler) {
-  this.reply_handler = reply_handler;
-  this.error_handler = error_handler;
-
-  this.send_message(this.msg);
-  this.state = 'PENDING';
-};
-
-// This takes a list of protocol definitions and dynamically generates methods on
-// the object that reflect that protocol.  These methods wrap ReplyCallback
-// objects which manage the reply and error callbacks of a remote proceedure call.
-// Remote defines a '_callbacks' variable which is a dict of message id's to
-// ReplyCallback objects.
-var Remote = function (transport, protocols) {
-  this.send_msg = transport;
-  this._callbacks = {};
-
-  _.each(protocols, function (protocol) {
-    this[protocol.procedure] = function (...args) {
-      var optionals = {};
-      // too few arguments
-      if (args.length < protocol.required.length) {
-        throw jsonrpc.InvalidRequest(
-          'Too few arguments passed to ' + protocol.procedure);
-      } else if (args.length > protocol.required.length + 1) {
-        throw jsonrpc.InvalidRequest(
-          'Too many arguments passed to ' + protocol.procedure);
-        // Options have been passed in (maybe)
-      } else if (args.length === protocol.required.length + 1) {
-        optionals = args.pop();
-        // optionals is not an object?
-        if (optionals === null || !_.isObject(optionals)) {
-          throw jsonrpc.InvalidRequest(
-            'Optional arguments must be an object, recieved ' +
-              optionals + ' for proceedure ' + protocol.proceedure);
-        }
-      } // else args.length == required.length and optionals should be empty
-
-      var params = _.zip(args, protocol.required).map(function ([a, p]) {
-        return {key: p.key, value: a, required: true};
-      }).concat(_.keys(optionals).map(function (k) {
-        return {key: k, value: optionals[k], required: false};
-      }));
-      // how to handle kwargs?
-      var msg = jsonrpc.request(protocol.procedure, params);
-      // Generating a UUID http://stackoverflow.com/a/2117523
-      msg.id = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
-        var r = Math.random() * 16 | 0;
-        var v = c === 'x' ? r : (r & 0x3 | 0x8);
-        return v.toString(16);
-      });
-
-      this._callbacks[msg.id] = new ReplyCallback(msg, this.send_msg);
-
-      return this._callbacks[msg.id];
-    }.bind(this);
-  }, this);
-};
-
-Remote.prototype.resolve = function (msg) {
-  if (msg.id in this._callbacks) {
-    // Resolve the callback
-    if (msg.error !== null) {
-      this._callbacks[msg.id].error_handler(msg.error);
-    } else {
-      this._callbacks[msg.id].reply_handler(msg.result);
-    }
-
-    delete this._callbacks[msg.id];
-  } else {
-    console.log("WARNING: Couldn't find callback for message id: " + msg.id); // eslint-disable-line no-console
-  }
-};
+import MapObject from './MapObject';
+import {
+  is_response,
+  is_request,
+  response,
+  constants,
+  Remote
+} from './jsonrpc';
 
 // the Geonotebook object manages the comm channel that is opened on
 // kernel initialization,  the 'Map' object,  which is a wrapper around
@@ -99,7 +18,7 @@ var Geonotebook = function (Jupyter, events) {
   this._remote = null;
 
   this.init_html_and_css();
-  this.map = new GeoMap(this);
+  this.map = new MapObject(this);
   this.register_events(Jupyter, events);
   this.load_annotation_buttons(Jupyter);
   Jupyter.map = this.map;
@@ -148,12 +67,12 @@ Geonotebook.prototype.resolve_arg_list = function (fn, msg) {
       } else if (arg['default']) {
         return undefined;
       } else {
-        throw jsonrpc.InvalidParam(
+        throw constants.InvalidParam(
           msg.method + ' did not recieve a required param ' + arg['key']);
       }
     });
   } else {
-    throw jsonrpc.ParseError('$arg_meta not available on ' + msg.method + '!');
+    throw constants.ParseError('$arg_meta not available on ' + msg.method + '!');
   }
 };
 
@@ -171,9 +90,9 @@ Geonotebook.prototype.recv_msg = function (message) {
     this.map.init_map();
   } else if (this.protocol_negotiation_complete) {
     // Pass response messages on to remote to be resolved
-    if (jsonrpc.is_response(msg)) {
+    if (is_response(msg)) {
       this._remote.resolve(msg);
-    } else if (jsonrpc.is_request(msg)) {
+    } else if (is_request(msg)) {
       try {
         // Apply the map method from the msg on the parameters
         var obj = this.get_object(msg);
@@ -182,23 +101,23 @@ Geonotebook.prototype.recv_msg = function (message) {
           var result = obj[msg.method].apply(
             this.map, this.resolve_arg_list(obj[msg.method], msg));
           // Reply with the result of the call
-          this.send_msg(jsonrpc.response(result, null, msg.id));
+          this.send_msg(response(result, null, msg.id));
         } else {
-          throw jsonrpc.MethodNotfound('Method ' + msg.method + ' not found!');
+          throw constants.MethodNotfound('Method ' + msg.method + ' not found!');
         }
       } catch (ex) {
         // If we catch an error report it back to the RPC caller
-        this.send_msg(jsonrpc.response(null, ex, msg.id));
+        this.send_msg(response(null, ex, msg.id));
       }
     } else {
       // Not a response or a request - send parse error
       this.send_msg(
-        jsonrpc.response(null, jsonrpc.ParseError('Could not parse message'), msg.id));
+        response(null, constants.ParseError('Could not parse message'), msg.id));
     }
   } else {
     // Protocol negotiation not complete - send internal error
     this.send_msg(
-      null, jsonrpc.InternalError('ERROR: Recieved a ' + msg.method + ' message ' +
+      null, constants.InternalError('ERROR: Recieved a ' + msg.method + ' message ' +
         'but protocol negotiation is not complete.'), msg.id);
   }
 };
