+++ js/src/index.js
@@ -1,5 +1,15 @@
+import './css/styles.css';
+
+// import all of geojs to get optional renderers
+// we could pick and choose what we want in the future
 import 'geojs';
 
-import geonotebook from './geonotebook';
+import Geonotebook from './Geonotebook';
+import MapObject from './MapObject';
+import * as jsonrpc from './jsonrpc';
 
-export default geonotebook;
+export {
+  Geonotebook,
+  MapObject,
+  jsonrpc
+};
