+++ js/src/jsonrpc/response.js
@@ -1,0 +1,8 @@
+function response (result, error, id) {
+        // TODO validate taht result/error are mutually exclusive
+  return {result: result,
+    error: error,
+    id: id };
+}
+
+export default response;
