--- js/src/utils/index.js
@@ -1,5 +1,0 @@
-import annotate from './annotate';
-
-export {
-  annotate
-};
